Status effects applied for an additional turn
The only way we can decrement the status effect turns (among other things) is to do so at the end of the BattlePhaseResult application. Thus, to ensure that Pokemon are affected by the status effect, we have to add a turn because it is applied then immediately decremented by a turn. E.g. if a Pokemon were set to sleep for 1 turn, they would get the effect, then it would be decremented in the same battle phase, then the effect would go away. This +1 is to ensure that the same-turn result application decrement step does not immediately remove the effect.